refactor(popover): split types in new modules
