+++ packages/lumx-core/src/js/components/Popover/constants.ts
@@ -27,21 +27,6 @@ export const Placement = {
 export type Placement = ValueOf<typeof Placement>;
 
 /**
- * Offset of the popover.
- */
-export interface Offset {
-    /** Offset size along the reference. */
-    along?: number;
-    /** Offset size away from the reference. */
-    away?: number;
-}
-
-/**
- * Popover elevation index.
- */
-export type Elevation = 1 | 2 | 3 | 4 | 5;
-
-/**
  * Popover fit anchor width options.
  */
 export const FitAnchorWidth = {
