+++ packages/lumx-core/src/js/components/Popover/types.ts
@@ -1,0 +1,14 @@
+/**
+ * Offset of the popover.
+ */
+export interface Offset {
+    /** Offset size along the reference. */
+    along?: number;
+    /** Offset size away from the reference. */
+    away?: number;
+}
+
+/**
+ * Popover elevation index.
+ */
+export type Elevation = 1 | 2 | 3 | 4 | 5;
